third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/css/css-position/position-absolute-in-inline-001-expected.txt

This is a testharness.js-based test.
FAIL absolute inside inline container location should be correct. assert_equals: expected 0 but got 100
Harness: the test ran to completion.
